feat(core): 重构工作线程并添加新功能
- 重构了 work_thread.js，优化了代码结构和逻辑
- 在 workspace.js 中实现了缓存目录的重用和清理机制
- 添加了 virtua 依赖用于前端组件
- 更新了组件类型定义，增加了 AModal
- 移除了 README.md 和 README.zh-CN.md 中的冗余信息
- 更新了 package.json 版本号 1.0.18

+++ client/electron/service/workspace.js
@@ -6,7 +6,6 @@ const ipcWorkMsg = 'controller.workspace.workMsg'
 const Ps = require('ee-core/ps')
 const fs = require('fs-extra')
 const path = require('path')
-const shortid = require('shortid')
 
 /**
  * 示例服务（service层为单例）
@@ -14,35 +13,41 @@ const shortid = require('shortid')
  */
 class WorkspaceService extends Service {
 
-	constructor(ctx) {
-		super(ctx)
-	}
+  constructor(ctx) {
+    super(ctx)
+  }
 
-	/**
-	 * test
-	 */
-	async run(args, event) {
-		const userDataDir = path.join(Ps.getExtraResourcesDir(), 'cache', shortid.generate())
-		const proc = fork(
-			require.resolve('../common/work_thread.js'),
-			[args.ua, args.ip, args.port, args.username, args.password, JSON.stringify(args.size), args.wr, args.wv, args.url, args.id, Ps.getExtraResourcesDir(), userDataDir, args.core], // pass to process.argv into child
-			{},
-		)
-		proc.on('message', (m) => {
-			// console.log(m)
-			event.sender.send(ipcWorkMsg, m)
-		})
-		proc.on('close', async (code) => {
-			console.log(code)
-			if (code === 0) {
-				event.sender.send(ipcWorkMsg, { id: args.id, type: 'end' })
-			} else {
-				event.sender.send(ipcWorkMsg, { id: args.id, type: 'error', message: '' })
-			}
-			await fs.remove(userDataDir)
-		})
-		return true
-	}
+  /**
+   * test
+   */
+  async run(args, event) {
+    // args.runCount, args.resweep
+    const userDataDir = path.join(Ps.getExtraResourcesDir(), 'cache', `cache-${args.id}-base`)
+    const proc = fork(
+      require.resolve('../common/work_thread.js'),
+      [args.ua, args.ip, args.port, args.username, args.password, JSON.stringify(args.size), args.wr, args.wv, args.url, args.id, Ps.getExtraResourcesDir(), userDataDir, args.core, args.random, args.randomLeft, args.randomRight, args.randomWEBGL], // pass to process.argv into child
+      {}
+    )
+    proc.on('message', (m) => {
+      // console.log(m)
+      event.sender.send(ipcWorkMsg, m)
+    })
+    proc.on('close', async (code) => {
+      if (code === 0) {
+        event.sender.send(ipcWorkMsg, { id: args.id, type: 'end' })
+      } else {
+        event.sender.send(ipcWorkMsg, { id: args.id, type: 'error', message: '' })
+      }
+      if (args.resweep === 0 || args.resweep === args.runCount) {
+        try {
+          await fs.remove(userDataDir)
+        } catch (error) {
+          console.error(`Failed to remove ${userDataDir}:`, error)
+        }
+      }
+    })
+    return true
+  }
 }
 
 WorkspaceService.toString = () => '[class WorkspaceService]'
